add some routes in backend
