+++ backend/src/controllers/app.controller.ts
@@ -1,0 +1,18 @@
+import { Controller, Get, Param } from '@nestjs/common';
+import { AppService } from '../services/app.service';
+
+@Controller()
+export class AppController {
+  constructor(private readonly appService: AppService) {}
+
+  @Get()
+  getHello(): string {
+    return this.appService.getHello();
+  }
+
+  @Get(':id')
+  getUserId(@Param('id') id: number): string {
+    console.log(id);
+    return `this request returns the user with id ${id}`;
+  }
+}
