+++ backend/src/controllers/index.ts
@@ -1,0 +1,2 @@
+export { AppController } from './app.controller';
+export { UserController } from './user.controller';
