+++ backend/src/modules/app.module.ts
@@ -2,13 +2,13 @@
 
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
-import databaseService from './config/database.service';
+import { AppController, UserController } from '../controllers/';
+import { AppService } from '../services/app.service';
+import databaseService from '../config/database.service';
 
 @Module({
   imports: [TypeOrmModule.forRoot(databaseService)],
-  controllers: [AppController],
+  controllers: [AppController, UserController],
   providers: [AppService],
 })
 export class AppModule {}
